+++ src/include/sas_export_function.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "duckdb.hpp"
+
+namespace duckdb {
+
+class ExtensionLoader;
+
+void RegisterSasExport(ExtensionLoader &loader);
+
+} // namespace duckdb